test wip

+++ src/Concerns/SyncsWithCampaignMonitorList.php
@@ -6,13 +6,16 @@ use BernskioldMedia\LaravelCampaignMonitor\Contracts\CampaignMonitorList;
 use BernskioldMedia\LaravelCampaignMonitor\Jobs\CreateCampaignMonitorList;
 use BernskioldMedia\LaravelCampaignMonitor\Jobs\DeleteCampaignMonitorList;
 use BernskioldMedia\LaravelCampaignMonitor\Jobs\UpdateCampaignMonitorList;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
+use function class_uses;
 use function dispatch;
+use function in_array;
 use function str;
 
 trait SyncsWithCampaignMonitorList
 {
-    public function bootSyncsWithCampaignMonitorList()
+    public static function bootSyncsWithCampaignMonitorList()
     {
         static::created(function (CampaignMonitorList $model) {
             if (! $model->shouldSyncWithCampaignMonitor()) {
@@ -30,13 +33,15 @@ trait SyncsWithCampaignMonitorList
             dispatch(new UpdateCampaignMonitorList($model));
         });
 
-        static::forceDeleted(function (CampaignMonitorList $model) {
-            if (! $model->shouldSyncWithCampaignMonitor()) {
-                return;
-            }
+        if (in_array(SoftDeletes::class, class_uses(static::class))) {
+            static::forceDeleted(function (CampaignMonitorList $model) {
+                if (! $model->shouldSyncWithCampaignMonitor()) {
+                    return;
+                }
 
-            dispatch(new DeleteCampaignMonitorList($model));
-        });
+                dispatch(new DeleteCampaignMonitorList($model));
+            });
+        }
     }
 
     public function getCampaignMonitorUniqueJobIdentifier(): string
